Refactor client registration types namespace
