+++ src/Codebooks/ClaimValues/ClientRegistrationTypesEnum.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace SimpleSAML\Module\oidc\Codebooks;
+namespace SimpleSAML\Module\oidc\Codebooks\ClaimValues;
 
 enum ClientRegistrationTypesEnum: string
 {
