+++ src/Forms/ClientForm.php
@@ -18,7 +18,7 @@ namespace SimpleSAML\Module\oidc\Forms;
 
 use Nette\Forms\Form;
 use SimpleSAML\Auth\Source;
-use SimpleSAML\Module\oidc\Codebooks\ClientRegistrationTypesEnum;
+use SimpleSAML\Module\oidc\Codebooks\ClaimValues\ClientRegistrationTypesEnum;
 use SimpleSAML\Module\oidc\Forms\Controls\CsrfProtection;
 use SimpleSAML\Module\oidc\ModuleConfig;
 use Traversable;
